+++ src/main/java/Models/RoomService.java
@@ -1,5 +1,6 @@
 package Models;
 
+import java.math.BigDecimal;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Date;
@@ -8,10 +9,10 @@ public class RoomService {
     private int id;
     private int reservation_id;
     private String service_name;
-    private double price;
+    private BigDecimal price;
     private Date requested_at;
 
-    public RoomService(int id, int reservation_id, String service_name, double price, Date requested_at){
+    public RoomService(int id, int reservation_id, String service_name, BigDecimal price, Date requested_at){
         this.id=id;
         this.reservation_id=reservation_id;
         this.service_name=service_name;
@@ -23,7 +24,7 @@ public class RoomService {
         int id = resultSet.getInt("id");
         int reservation_id=resultSet.getInt("reservation_id");
         String service_name=resultSet.getString("service_name");
-        double price=resultSet.getDouble("price");
+        BigDecimal price=resultSet.getBigDecimal("price");
         Date requested_at=resultSet.getDate("requested_at");
 
         return new RoomService(id,reservation_id,service_name,price,requested_at);
@@ -35,7 +36,7 @@ public class RoomService {
 
     public String getService_name() { return service_name; }
 
-    public double getPrice() { return price; }
+    public BigDecimal getPrice() { return price; }
 
     public Date getRequested_at() { return requested_at; }
 }
